// src/types/StockEntry.type.ts
import { StockEntryItem } from "./StockEntryItem.type";

export interface StockEntry {
	name?: string;
	owner?: string;
	creation?: string;
	modified?: string;
	modified_by?: string;
	docstatus?: number;
	idx?: number;
	workflow_state?: string;
	naming_series?: string;
	stock_entry_type?: string;
	outgoing_stock_entry?: string;
	purpose?: string;
	add_to_transit?: number;
	work_order?: string;
	purchase_order?: string;
	subcontracting_order?: string;
	delivery_note_no?: string;
	sales_invoice_no?: string;
	pick_list?: string;
	purchase_receipt_no?: string;
	company?: string;
	posting_date?: string;
	posting_time?: string;
	set_posting_time?: number;
	inspection_required?: number;
	from_bom?: number;
	apply_putaway_rule?: number;
	bom_no?: string;
	fg_completed_qty?: number;
	use_multi_level_bom?: number;
	from_warehouse?: string;
	source_warehouse_address?: string;
	source_address_display?: string;
	to_warehouse?: string;
	target_warehouse_address?: string;
	target_address_display?: string;
	scan_barcode?: string;
	total_outgoing_value?: number;
	total_incoming_value?: number;
	value_difference?: number;
	total_additional_costs?: number;
	supplier?: string;
	supplier_name?: string;
	supplier_address?: string;
	address_display?: string;
	project?: string;
	select_print_heading?: string;
	letter_head?: string;
	is_opening?: string;
	remarks?: string;
	per_transferred?: number;
	total_amount?: number;
	job_card?: string;
	amended_from?: string;
	credit_note?: string;
	is_return?: number;
	doctype?: string;
	items?: StockEntryItem[];
	additional_costs?: any[];
	empty_truck_weight?: number;
	total_weight_difference_with_truck_weight?: number;
	full_truck_weight?: number;
	total_weight?: number;
	driver_name?: string;
	driver_phone_number?: string;
	bonnet?: string;
	harvest_method?: string;
	complementary_info1?: string;
	complementary_info2?: string;
	vehicle_number?: string;
}